+++ Structural/Flyweight/src/Interfaces/Product.ts
@@ -1,0 +1,6 @@
+export interface Product {
+  id: number;
+  name: string;
+  price: number;
+  imageSrc: string;
+}
