+++ Structural/Flyweight/src/models/ImageFactory.ts
@@ -1,0 +1,12 @@
+export class ImageFactory {
+  private imageCache: Map<string, HTMLImageElement> = new Map();
+
+  getImage(src: string): HTMLImageElement {
+    if (!this.imageCache.has(src)) {
+      const img = new Image();
+      img.src = src;
+      this.imageCache.set(src, img);
+    }
+    return this.imageCache.get(src)!;
+  }
+}
