+++ Structural/Proxy/src/interfaces/DatabaseAccess.ts
@@ -1,0 +1,3 @@
+export interface DatabaseAccess {
+  writeData(): void;
+}
