+++ Structural/Proxy/src/models/User.ts
@@ -1,0 +1,11 @@
+export class User {
+  constructor(private _writeAccess: boolean) {}
+
+  get writeAccess() {
+    return this._writeAccess;
+  }
+
+  set writeAccess(access: boolean) {
+    this._writeAccess = access;
+  }
+}
