--- Structural/Proxy/src/proxyPattern.ts
@@ -1,90 +1,0 @@
-const form = document.querySelector('form');
-form?.addEventListener('submit', (e) => {
-    e.preventDefault();
-})
-const databaseList = document.querySelector('.database-list');
-const databaseContent = document.querySelector('.database-content');
-const warningContent = document.querySelector('.warning-content');
-const writeAccessButton = document.querySelector('.write-access');
-const writeDatabaseInput = <HTMLInputElement>document.querySelector('.write-database');
-
-function insertDatabase(data: string) {
-    let li = document.createElement("li");
-    li.appendChild(document.createTextNode(data));
-    databaseList?.appendChild(li);
-}
-
-interface DatabaseAccess {
-    writeData() : void;
-}
-
-class RealDatabaseAccess implements DatabaseAccess {
-    private data: string;
-
-    constructor(data: string) {
-        this.data = data;
-    }
-
-    writeData() {
-        insertDatabase(this.data);
-        console.log("Real write into database");
-    } 
-}
-
-class ProtectionProxyDatabaseAccess implements DatabaseAccess {
-    private realDatabaseAccess : DatabaseAccess | null = null;
-
-    constructor(private user: User, private data: string) {
-        if (user.writeAccess) {
-            this.realDatabaseAccess = new RealDatabaseAccess(data);
-            console.log("User has write access");
-        } else {
-            console.log("User does not have write access");
-        }
-    }
-
-    writeData() {
-        if (this.realDatabaseAccess && this.user.writeAccess) {
-            this.realDatabaseAccess.writeData();
-            console.log("Proxy write into database");
-        } 
-    }
-}
-
-class User {
-    constructor(private _writeAccess : boolean) {}
-
-    get writeAccess() {
-        return this._writeAccess;
-    }
-
-    set writeAccess(access: boolean) {
-        this._writeAccess = access;
-    }
-}
-
-let writeAccess = false;
-const user = new User(writeAccess);
-
-writeAccessButton?.addEventListener('click', () => {
-    if (!user.writeAccess) {
-        writeAccessButton?.classList.remove("deny");
-        writeAccessButton?.classList.add("allow");
-        writeDatabaseInput?.classList.remove("deny");
-        user.writeAccess = true;
-    } else {
-        writeAccessButton?.classList.remove("allow");
-        writeAccessButton?.classList.add("deny");
-        writeDatabaseInput?.classList.add("deny");
-        user.writeAccess = false;
-    }
-});
-
-function addDatabase() {
-    const form = document.querySelector('form')!;
-    const elements = form.elements!;
-    const data = (elements[0] as HTMLFormElement).value.toString();
-    writeDatabaseInput.value = "";
-    const proxyDatabase = new ProtectionProxyDatabaseAccess(user, data);
-    proxyDatabase.writeData();
-}
